restructured project: now there are several CMakeLists.txt files, one in each source directory gtest setup WIP: currently not compiling tests updated dependency graph and todo.md

+++ include/utils.h
@@ -63,7 +63,7 @@ typedef struct {
 /* types used for process communication */
 typedef struct {
     /* common header */
-    uint32_t id;                /* transaction ID */
+    uint16_t id;                /* transaction ID */
     char op_code;               /* operation code that indicates the client API function called */
 } header_t;
